- use double type for better precision (64 bit)

+++ src/asteroid.h
@@ -14,7 +14,7 @@
  * @param albedo Reflective power of the asteroid
  * @return diameter of the asteroid in meters
  */
-float ac_asteroid_diameter(float abs_mag, float albedo);
+double ac_asteroid_diameter(double abs_mag, double albedo);
 
 /**
  * @brief Calculate the apparent diameter of an asteroid
@@ -22,6 +22,6 @@ float ac_asteroid_diameter(float abs_mag, float albedo);
  * @param asteroid_earth_dist Asteroid to Earth distance in Astronomical Unit (AU)
  * @return apparent diameter of the asteroid in meters
  */
-float ac_asteroid_apparent_diameter(float true_diameter, float asteroid_earth_dist);
+double ac_asteroid_apparent_diameter(double true_diameter, double asteroid_earth_dist);
 
 #endif
